+++ Packages/BrandonUtils/Standalone/Clerical/DuplicateFileResolution.cs
@@ -1,0 +1,7 @@
+﻿namespace BrandonUtils.Standalone.Clerical {
+    public enum DuplicateFileResolution {
+        Error = default,
+        Overwrite,
+        Backup,
+    }
+}